refactor(chat): move thread scroll anchoring into a hook
ChatThreadView carried a three-mode scroll state machine inside the
render component: nine refs, seven effects and the virtualizer. It now
lives in useChatScrollAnchor with the component tests as its contract.
The message list is a polite live region so a reply is announced, the
execution view no longer nests an li in a div, the dead meta layout is
gone, and the thread styles use tokens instead of em and px literals.
